// tamkin-backend/src/Common/Enums/Jwt/jwt.enum.ts
export enum DeviceTypeEnum {
  DESKTOP = 'Desktop',
  MOBILE = 'Mobile',
  TABLET = 'Tablet',
}

export enum OSEnum {
  LINUX = 'Linux',
  WINDOWS = 'Windows',
  MAC_OS = 'Mac Os',
  ANDROID = 'Android',
  IOS = 'Ios',
}

export enum BrowserEnum {
  CHROME = 'Chrome',
  FIREFOX = 'Firefox',
  SAFARI = 'Safari',
  EDGE = 'Edge',
}